Reference

LocalStore.clean — Method

clean(obj)

Removes the stored data associated with obj.

LocalStore.clean — Method

clean(; force=false)

Removes all stored data directories.

LocalStore.expected_tree_hash — Method

expected_tree_hash(obj)

Expected tree hash for the saved contents of obj. Overload this for specific item types. Otherwise simply returns the hash of the contents of the store (so no verification happens).

LocalStore.issaved — Method

issaved(obj)

Returns true if obj is locally stored, false otherwise.

LocalStore.itemdir — Method

itemdir(obj)

Directory path where obj will be stored.

LocalStore.load — Method

load(obj)

Reads and returns the specified object. Stores it locally if it hasn't been stored already.

LocalStore.rootdir — Method

rootdir()

Root directory path where objects will be stored.

LocalStore.save — Method

save(obj)

Stores obj locally if it has not been stored already. Returns path to directory containing obj.

LocalStore.verify — Method

verify(obj)

Compares the treehash of the stored copy of obj to the expected treehash, and errors if there is a difference.